Removed a bug related to malloc(0). When the argument to the malloc is zero the pointer is non-null and the deallocation is done even though we check if (data).. Then the deallocation is tried on a zero sized array.

+++ MandleBrot/MandelBrot/MandelBrot/Algorithm/mandelbrot.cpp
@@ -19,7 +19,7 @@ mandelbrot::mandelbrot(const allocation_mode& _alloc_mode, const allocation_majo
 	file_name{ _file_name }
 {
 	array_alloc_ptr = std::make_unique<array_allocator>(alloc_mode, alloc_major, n_xs, n_ys, file_name,gamma_);
-	this->area = 0.0;
+	this->area = 0;
 }
 
 
@@ -56,7 +56,8 @@ void mandelbrot::output()
 {
 	array_alloc_ptr->write_data(n_xs,n_ys);
 }
-double& mandelbrot::return_area()
+
+int& mandelbrot::return_area()
 {
 	return this->area;
 }
